Add verified_type field to User schema
This commit adds a new field to the User schema that indicates the type of verification for the Twitter account. The field can take one of four values: blue, business, government, or none. To retrieve the verified_type field in API requests, users should include the user.fields=verified_type parameter in the query string.

Signed-off-by: LMSPaul <114345821+LMSPaul@users.noreply.github.com>